Get rid of translation layer
This commit restructures the libdrop changes so that the big chunk of
code related to translating data representaton between libdrop and
internal data structures is not needed anymore.
