+++ fileshare/libdrop/libdrop.go
@@ -5,12 +5,16 @@ import (
 	"fmt"
 	"log"
 	"net/netip"
+	"path/filepath"
+	"strings"
 	"sync"
 	"time"
 
 	norddrop "github.com/NordSecurity/libdrop-go/v7"
 	"github.com/NordSecurity/nordvpn-linux/fileshare"
+	"github.com/NordSecurity/nordvpn-linux/fileshare/pb"
 	"github.com/NordSecurity/nordvpn-linux/internal"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 // Fileshare is the main functional filesharing implementation using norddrop library.
@@ -315,21 +319,20 @@ func (f *Fileshare) CancelFile(transferID string, fileID string) error {
 	return nil
 }
 
-// GetTransfersSince provided time from fileshare implementation storage
-func (f *Fileshare) GetTransfersSince(t time.Time) ([]fileshare.LibdropTransfer, error) {
+// Load transfers from fileshare implementation storage
+func (f *Fileshare) Load() (map[string]*pb.Transfer, error) {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
-	since := t.Unix()
+	since := time.Time{}.Unix()
 	norddropTransfers, err := f.norddrop.TransfersSince(since)
 	if err != nil {
 		return nil, fmt.Errorf("getting transfers since %d: %w", since, err)
 	}
 
-	transfers := make([]fileshare.LibdropTransfer, len(norddropTransfers))
-
-	for i, transfer := range norddropTransfers {
-		transfers[i] = toInternalTransfer(&transfer)
+	transfers := make(map[string]*pb.Transfer)
+	for _, transfer := range norddropTransfers {
+		transfers[transfer.Id] = norddropTransferToPBTransfer(transfer)
 	}
 
 	return transfers, nil
@@ -348,180 +351,359 @@ func (f *Fileshare) PurgeTransfersUntil(until time.Time) error {
 	return f.norddrop.PurgeTransfersUntil(until.Unix() * 100)
 }
 
-func toInternalTransfer(ti *norddrop.TransferInfo) fileshare.LibdropTransfer {
-	return fileshare.LibdropTransfer{
-		Kind:      toInternalTransferKind(ti.Kind),
-		Id:        ti.Id,
-		Peer:      ti.Peer,
-		States:    toInternalTransferStates(ti.States),
-		CreatedAt: ti.CreatedAt,
+func norddropTransferToPBTransfer(ti norddrop.TransferInfo) *pb.Transfer {
+	allFiles := filesFromTransferInfo(&ti)
+	totalSize, totalTransferred := calculateTotalSizeAndTotalTransferred(allFiles)
+	out := &pb.Transfer{
+		Id:               ti.Id,
+		Direction:        directionFromTransferInfo(&ti),
+		Peer:             ti.Peer,
+		Created:          timestamppb.New(time.UnixMilli(ti.CreatedAt)),
+		Files:            allFiles,
+		TotalSize:        totalSize,
+		TotalTransferred: totalTransferred,
+		Status:           determineTransferStatusAndAdjustFileStatuses(&ti, allFiles),
+		Path:             determineTransferPath(&ti, allFiles),
 	}
+
+	return out
 }
 
-func toInternalTransferKind(transferKind norddrop.TransferKind) fileshare.TransferKind {
-	switch v := transferKind.(type) {
-	case norddrop.TransferKindIncoming:
-		return fileshare.TransferKindIncoming{
-			Paths: toInternalIncomingPaths(v.Paths),
-		}
+func directionFromTransferInfo(ti *norddrop.TransferInfo) pb.Direction {
+	var direction pb.Direction
+	switch ti.Kind.(type) {
 	case norddrop.TransferKindOutgoing:
-		return fileshare.TransferKindOutgoing{
-			Paths: toInternalOutgoingPaths(v.Paths),
-		}
+		direction = pb.Direction_OUTGOING
+	case norddrop.TransferKindIncoming:
+		direction = pb.Direction_INCOMING
 	default:
-		log.Printf(internal.WarningPrefix+" unexpected norddrop.TransferKind: %T\n", v)
-		return fileshare.TransferKindUnknown{}
+		log.Printf(internal.WarningPrefix+" unknown direction found when parsing libdrop transfers: %T\n", ti.Kind)
+		direction = pb.Direction_UNKNOWN_DIRECTION
 	}
+	return direction
 }
 
-func toInternalIncomingPaths(paths []norddrop.IncomingPath) []fileshare.IncomingPath {
-	result := make([]fileshare.IncomingPath, len(paths))
-	for i, path := range paths {
-		result[i] = fileshare.IncomingPath{
-			FileId:        path.FileId,
-			RelativePath:  path.RelativePath,
-			Bytes:         path.Bytes,
-			BytesReceived: path.BytesReceived,
-			States:        toInternalIncomingPathStates(path.States),
+func filesFromTransferInfo(ri *norddrop.TransferInfo) []*pb.File {
+	switch ti := ri.Kind.(type) {
+	case norddrop.TransferKindIncoming:
+		files := make([]*pb.File, len(ti.Paths))
+		for i, path := range ti.Paths {
+			files[i] = norddropIncomingPathToPBFile(path)
 		}
-	}
-	return result
-}
-
-func toInternalIncomingPathStates(states []norddrop.IncomingPathState) []fileshare.IncomingPathState {
-	result := make([]fileshare.IncomingPathState, len(states))
-	for i, state := range states {
-		result[i] = fileshare.IncomingPathState{
-			Kind:      toInternalIncomingPathStateKind(state.Kind),
-			CreatedAt: state.CreatedAt,
+		return files
+	case norddrop.TransferKindOutgoing:
+		files := make([]*pb.File, len(ti.Paths))
+		for i, path := range ti.Paths {
+			files[i] = norddropOutgoingPathToPBFile(path)
 		}
-	}
-	return result
-}
-
-func toInternalIncomingPathStateKind(kind norddrop.IncomingPathStateKind) fileshare.IncomingPathStateKind {
-	switch v := kind.(type) {
-	case norddrop.IncomingPathStateKindCompleted:
-		return fileshare.IncomingPathStateKindCompleted{
-			FinalPath: v.FinalPath,
-		}
-	case norddrop.IncomingPathStateKindFailed:
-		return fileshare.IncomingPathStateKindFailed{
-			Status:        fileshare.StatusCode(v.Status),
-			BytesReceived: v.BytesReceived,
-		}
-	case norddrop.IncomingPathStateKindPaused:
-		return fileshare.IncomingPathStateKindPaused{
-			BytesReceived: v.BytesReceived,
-		}
-	case norddrop.IncomingPathStateKindPending:
-		return fileshare.IncomingPathStateKindPending{
-			BaseDir: v.BaseDir,
-		}
-	case norddrop.IncomingPathStateKindRejected:
-		return fileshare.IncomingPathStateKindRejected{
-			ByPeer:        v.ByPeer,
-			BytesReceived: v.BytesReceived,
-		}
-	case norddrop.IncomingPathStateKindStarted:
-		return fileshare.IncomingPathStateKindStarted{
-			BytesReceived: v.BytesReceived,
-		}
+		return files
 	default:
-		log.Printf(internal.WarningPrefix+" unexpected norddrop.IncomingPathStateKind: %T\n", v)
-		return fileshare.IncomingPathStateKindUnknown{}
+		log.Printf(internal.WarningPrefix+" unknown transfer kind: %T\n", ti)
+		return []*pb.File{}
 	}
 }
 
-func toInternalTransferStates(transferStates []norddrop.TransferState) []fileshare.TransferState {
-	states := make([]fileshare.TransferState, len(transferStates))
-	for i, state := range transferStates {
-		states[i] = fileshare.TransferState{
-			Kind:      toInternalTransferStateKind(state.Kind),
-			CreatedAt: state.CreatedAt,
+func norddropOutgoingPathToPBFile(outPath norddrop.OutgoingPath) *pb.File {
+	file := &pb.File{
+		Id:     outPath.FileId,
+		Path:   outPath.RelativePath,
+		Size:   outPath.Bytes,
+		Status: statusFromOutgoingPath(&outPath),
+		// This only shows the amount from the last status change.
+		// The only up to date source for data transferred are events.
+		// To show the correct value we track ongoing transfers in event manager,
+		// and event manager overwrites values of transferred fields if needed.
+		Transferred: transferredFromOutgoingPath(&outPath),
+		FullPath:    determineFullOutgoingPath(&outPath),
+	}
+	return file
+}
+
+func statusFromOutgoingPath(outPath *norddrop.OutgoingPath) pb.Status {
+	var status pb.Status
+
+	if len(outPath.States) == 0 {
+		status = pb.Status_REQUESTED
+	} else {
+		lastState := outPath.States[len(outPath.States)-1]
+		switch lastState := lastState.Kind.(type) {
+		case norddrop.OutgoingPathStateKindStarted:
+			status = pb.Status_ONGOING
+		case norddrop.OutgoingPathStateKindFailed:
+			status = pb.Status(lastState.Status)
+		case norddrop.OutgoingPathStateKindCompleted:
+			status = pb.Status_SUCCESS
+		case norddrop.OutgoingPathStateKindRejected:
+			status = pb.Status_CANCELED
+		case norddrop.OutgoingPathStateKindPaused:
+			status = pb.Status_PAUSED
+		default:
+			log.Printf(internal.WarningPrefix+" unknown file status in transfer: %T\n", lastState)
+			status = pb.Status_BAD_STATUS
 		}
 	}
-	return states
+	return status
 }
 
-func toInternalTransferStateKind(kind norddrop.TransferStateKind) fileshare.TransferStateKind {
-  switch v := kind.(type) {
-	case norddrop.TransferStateKindCancel:
-  return fileshare.TransferStateKindCancel{
-  	ByPeer: v.ByPeer,
-  }
-	case norddrop.TransferStateKindFailed:
-  return fileshare.TransferStateKindFailed{
-  	Status: fileshare.StatusCode(v.Status),
-  }
-	default:
-    log.Printf(internal.WarningPrefix+" unexpected norddrop.TransferStateKind: %T\n", v)
-  return fileshare.TransferStateKindUnknown{}
-	}
-}
-
-func toInternalOutgoingPaths(paths []norddrop.OutgoingPath) []fileshare.OutgoingPath {
-	result := make([]fileshare.OutgoingPath, len(paths))
-	for i, path := range paths {
-		result[i] = fileshare.OutgoingPath{
-			Source:       toInternalPathSource(path.Source),
-			FileId:       path.FileId,
-			RelativePath: path.RelativePath,
-			States:       toInternalOutgoingPathStates(path.States),
-			Bytes:        path.Bytes,
-			BytesSent:    path.BytesSent,
+func transferredFromOutgoingPath(outPath *norddrop.OutgoingPath) uint64 {
+	var transferred uint64
+	for _, state := range outPath.States {
+		switch st := state.Kind.(type) {
+		case norddrop.OutgoingPathStateKindStarted:
+			transferred = st.BytesSent
+		case norddrop.OutgoingPathStateKindCompleted:
+			transferred = outPath.Bytes
+		case norddrop.OutgoingPathStateKindFailed:
+			transferred = st.BytesSent
+		case norddrop.OutgoingPathStateKindRejected:
+			transferred = st.BytesSent
+		case norddrop.OutgoingPathStateKindPaused:
+			transferred = st.BytesSent
 		}
 	}
-	return result
+	return transferred
 }
 
-func toInternalPathSource(source norddrop.OutgoingFileSource) fileshare.OutgoingFileSource {
-	switch v := source.(type) {
+func determineFullOutgoingPath(outPath *norddrop.OutgoingPath) string {
+	switch pathSource := outPath.Source.(type) {
 	case norddrop.OutgoingFileSourceBasePath:
-		return fileshare.OutgoingFileSourceBasePath{
-			BasePath: v.BasePath,
-		}
+		return filepath.Join(pathSource.BasePath, outPath.RelativePath)
 	default:
-		log.Printf(internal.WarningPrefix+" unexpected norddrop.OutgoingFileSource: %T\n", v)
-		return fileshare.OutgoingFileSourceUnknown{}
+		log.Printf(internal.WarningPrefix+" unsupported path source: %T\n", outPath.Source)
+		return ""
 	}
 }
 
-func toInternalOutgoingPathStates(states []norddrop.OutgoingPathState) []fileshare.OutgoingPathState {
-	result := make([]fileshare.OutgoingPathState, len(states))
-	for i, state := range states {
-		result[i] = fileshare.OutgoingPathState{
-			Kind:      toInternalPathStateKind(state.Kind),
-			CreatedAt: state.CreatedAt,
+func norddropIncomingPathToPBFile(inPath norddrop.IncomingPath) *pb.File {
+	out := &pb.File{
+		Id:     inPath.FileId,
+		Path:   inPath.RelativePath,
+		Size:   inPath.Bytes,
+		Status: statusFromIncomingPath(&inPath),
+		// This only shows the amount from the last status change.
+		// The only up to date source for data transferred are events.
+		// To show the correct value we track ongoing transfers in event manager,
+		// and event manager overwrites values of transferred fields if needed.
+		Transferred: transferredFromIncomingPath(&inPath),
+		FullPath:    determineFullIncomingPath(&inPath),
+	}
+	return out
+}
+
+func statusFromIncomingPath(inPath *norddrop.IncomingPath) pb.Status {
+	var status pb.Status
+	if len(inPath.States) == 0 {
+		status = pb.Status_REQUESTED
+	} else {
+		lastState := inPath.States[len(inPath.States)-1]
+		switch lastState := lastState.Kind.(type) {
+		case norddrop.IncomingPathStateKindCompleted:
+			status = pb.Status_SUCCESS
+		case norddrop.IncomingPathStateKindFailed:
+			status = pb.Status(lastState.Status)
+		case norddrop.IncomingPathStateKindPaused:
+			status = pb.Status_PAUSED
+		case norddrop.IncomingPathStateKindPending:
+			status = pb.Status_PENDING
+		case norddrop.IncomingPathStateKindRejected:
+			status = pb.Status_CANCELED
+		case norddrop.IncomingPathStateKindStarted:
+			status = pb.Status_ONGOING
+		default:
+			log.Printf(internal.WarningPrefix+" unknown file status in transfer: %T\n", lastState)
+			status = pb.Status_BAD_STATUS
 		}
 	}
+	return status
+}
+
+func transferredFromIncomingPath(inPath *norddrop.IncomingPath) uint64 {
+	var transferred uint64
+	for _, state := range inPath.States {
+		switch st := state.Kind.(type) {
+		case norddrop.IncomingPathStateKindStarted:
+			transferred = st.BytesReceived
+		case norddrop.IncomingPathStateKindCompleted:
+			transferred = inPath.Bytes
+		case norddrop.IncomingPathStateKindFailed:
+			transferred = st.BytesReceived
+		case norddrop.IncomingPathStateKindRejected:
+			transferred = st.BytesReceived
+		case norddrop.IncomingPathStateKindPaused:
+			transferred = st.BytesReceived
+		}
+	}
+	return transferred
+}
+
+func determineFullIncomingPath(inPath *norddrop.IncomingPath) string {
+	fullPath := inPath.RelativePath
+	for _, state := range inPath.States {
+		switch st := state.Kind.(type) {
+		case norddrop.IncomingPathStateKindPending:
+			// BasePath is provided with the very first "pending" state
+			fullPath = filepath.Join(st.BaseDir, fullPath)
+		case norddrop.IncomingPathStateKindCompleted:
+			// FinalPath is provided with the very last "completed" state, so if it is present, then
+			// it will overwrite the previously constructed path
+			fullPath = st.FinalPath
+		}
+	}
+	return fullPath
+}
+
+func calculateTotalSizeAndTotalTransferred(allFiles []*pb.File) (uint64, uint64) {
+	var totalSize, totalTransferred uint64
+	for _, file := range allFiles {
+		if isFileTransferred(file) {
+			totalSize += file.Size
+			totalTransferred += file.Transferred
+		}
+	}
+	return totalSize, totalTransferred
+}
+
+func determineTransferStatusAndAdjustFileStatuses(libdropTransfer *norddrop.TransferInfo, allFiles []*pb.File) pb.Status {
+	var status pb.Status
+	if len(libdropTransfer.States) != 0 {
+		lastState := libdropTransfer.States[len(libdropTransfer.States)-1]
+		switch lastState := lastState.Kind.(type) {
+		case norddrop.TransferStateKindCancel:
+			// This is annoying. We have to "finalize" finished transfers by cancelling them,
+			// otherwise there's a resource leak in libdrop. Also, by doing that all finished
+			// transfers have cancel status, so we need to figure out the real status.
+			// So we try to determine status by files and see if the transfer was already finished
+			// when it was cancelled or not.
+			statusByFiles := getTransferStatusByFiles(allFiles)
+			if statusByFiles == pb.Status_SUCCESS || statusByFiles == pb.Status_FINISHED_WITH_ERRORS {
+				status = statusByFiles
+			} else {
+				if lastState.ByPeer {
+					status = pb.Status_CANCELED_BY_PEER
+				} else {
+					status = pb.Status_CANCELED
+				}
+				// adjust file statuses
+				for _, file := range allFiles {
+					if !isFileCompleted(file) {
+						file.Status = pb.Status_CANCELED
+					}
+				}
+			}
+		case norddrop.TransferStateKindFailed:
+			status = pb.Status(lastState.Status)
+		}
+	} else {
+		status = getTransferStatusByFiles(allFiles)
+	}
+	return status
+}
+
+func getTransferStatusByFiles(files []*pb.File) pb.Status {
+	allCanceled := true
+	allFinished := true
+	hasNoErrors := true
+	hasStarted := false
+	for _, file := range files {
+		if allCanceled && file.Status != pb.Status_CANCELED {
+			allCanceled = false
+		}
+		if allFinished && !isFileCompleted(file) {
+			allFinished = false
+		}
+		if hasNoErrors && checkFileHasErrors(file) {
+			hasNoErrors = false
+		}
+		if !hasStarted && file.Status != pb.Status_REQUESTED {
+			hasStarted = true
+		}
+	}
+
+	switch {
+	case allCanceled:
+		return pb.Status_CANCELED
+	case allFinished && hasNoErrors:
+		return pb.Status_SUCCESS
+	case allFinished && !hasNoErrors:
+		return pb.Status_FINISHED_WITH_ERRORS
+	case hasStarted:
+		return pb.Status_ONGOING
+	default:
+		return pb.Status_REQUESTED
+	}
+}
+
+func determineTransferPath(libdropTransfer *norddrop.TransferInfo, allFiles []*pb.File) string {
+	var result string
+	var transferPath string
+	// Base path is in different place for incoming and outgoing files
+	switch ti := libdropTransfer.Kind.(type) {
+	case norddrop.TransferKindIncoming:
+		for _, path := range ti.Paths {
+			for _, state := range path.States {
+				switch st := state.Kind.(type) {
+				case norddrop.IncomingPathStateKindPending:
+					transferPath = st.BaseDir
+				}
+			}
+
+			if result == "" {
+				result = transferPath
+			} else if result != transferPath && transferPath != "" {
+				result = "multiple files"
+			}
+		}
+		return transferPath
+	case norddrop.TransferKindOutgoing:
+		for i, path := range ti.Paths {
+			if filepath.Base(path.RelativePath) == path.RelativePath {
+				file := allFiles[i]
+				// If relative path doesn't contain a directory - means user specified a single file
+				// We show the full path in that case for outgoing transfers
+				transferPath = file.FullPath
+			} else {
+				// If user is sending a directory - we add the directory itself to the path.
+				// Example: user sends /tmp/test, then file.BasePath would be "/tmp", while
+				// file.RelativePath would be "test/file". So we take the dir name from RelativePath
+				// and add it to BasePath to get "/tmp/test".
+				dir, _, ok := strings.Cut(path.RelativePath, string(filepath.Separator))
+				if ok {
+					switch source := path.Source.(type) {
+					case norddrop.OutgoingFileSourceBasePath:
+						transferPath = filepath.Join(source.BasePath, dir)
+					default:
+						log.Printf(internal.WarningPrefix+" unsupported path source: %T\n", source)
+					}
+				}
+			}
+
+			if result == "" {
+				result = transferPath
+			} else if result != transferPath && transferPath != "" {
+				result = "multiple files"
+			}
+		}
+	}
+
 	return result
 }
 
-func toInternalPathStateKind(kind norddrop.OutgoingPathStateKind) fileshare.OutgoingPathStateKind {
-	switch v := kind.(type) {
-	case norddrop.OutgoingPathStateKindCompleted:
-		return fileshare.OutgoingPathStateKindCompleted{}
-	case norddrop.OutgoingPathStateKindFailed:
-		return fileshare.OutgoingPathStateKindFailed{
-			Status:    fileshare.StatusCode(v.Status),
-			BytesSent: v.BytesSent,
-		}
-	case norddrop.OutgoingPathStateKindPaused:
-		return fileshare.OutgoingPathStateKindPaused{
-			BytesSent: v.BytesSent,
-		}
-	case norddrop.OutgoingPathStateKindRejected:
-		return fileshare.OutgoingPathStateKindRejected{
-			ByPeer:    v.ByPeer,
-			BytesSent: v.BytesSent,
-		}
-	case norddrop.OutgoingPathStateKindStarted:
-		return fileshare.OutgoingPathStateKindStarted{
-			BytesSent: v.BytesSent,
-		}
-	default:
-		log.Printf(internal.WarningPrefix+" unexpected norddrop.OutgoingPathStateKind: %T\n", v)
-		return fileshare.OutgoingPathStateKindUnknown{}
-	}
+func checkFileHasErrors(file *pb.File) bool {
+	return file.Status != pb.Status_SUCCESS &&
+		file.Status != pb.Status_REQUESTED &&
+		file.Status != pb.Status_CANCELED &&
+		file.Status != pb.Status_ONGOING &&
+		file.Status != pb.Status_PENDING
+}
+
+func isFileCompleted(file *pb.File) bool {
+	return file.Status != pb.Status_REQUESTED &&
+		file.Status != pb.Status_ONGOING &&
+		file.Status != pb.Status_PENDING
+}
+
+// Used to check if file's size should be part of transfer's total size
+// Basically we don't include files that are canceled or errored out
+func isFileTransferred(file *pb.File) bool {
+	return !isFileCompleted(file) || file.Status == pb.Status_SUCCESS
 }
